chore: fixed express-ui, Refactored router implementation, removed unused code and updated dependencies
